pyproject.toml
[build-system]
requires = ["setuptools", "wheel", "cython>=3.2.0", "numpy>=1.26.4", "packaging"]

[project]
name = "matcha-tts"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "conformer==0.3.2",
    "cython>=3.2.0",
    "diffusers>=0.35.2",
    "einops>=0.8.1",
    "gdown>=5.2.0",
    "gradio==3.43.2",
    "hydra-colorlog==1.2.0",
    "hydra-core==1.3.2",
    "hydra-optuna-sweeper==1.2.0",
    "inflect>=7.5.0",
    "ipywidgets>=8.1.8",
    "librosa>=0.11.0",
    "matplotlib>=3.10.7",
    "notebook>=7.4.7",
    "numpy>=1.26.4",
    "pandas>=2.3.3",
    "peft>=0.18.0",
    "phonemizer>=3.3.0",
    "pre-commit>=4.3.0",
    "pytest>=8.4.2",
    "rich>=14.2.0",
    "rootutils>=1.0.7",
    "scipy>=1.16.3",
    "seaborn>=0.13.2",
    "tensorboard>=2.20.0",
    "torch>=2.0.0",
    "torchaudio>=2.9.0",
    "torchcodec>=0.9.1",
    "torchmetrics>=0.11.4",
    "torchvision>=0.15.0",
    "tqdm>=4.66.0",
    "unidecode>=1.4.0",
    "wget>=3.2",
]
